changed marker size #2432

+++ src/UI/Shared/WB.UI.Shared.Extensions/ViewModels/InterviewerMapDashboardViewModel.cs
@@ -43,7 +43,7 @@ public class InterviewerMapDashboardViewModel : MapDashboardViewModel
 
     public override bool SupportDifferentResponsible => false;
 
-    protected override Symbol GetInterviewMarkerSymbol(IInterviewMarkerViewModel interview, int size = 1)
+    protected override Symbol GetInterviewMarkerSymbol(IInterviewMarkerViewModel interview, double size = 1)
     {
         Color markerColor;
